ai-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
httpx==0.25.2
python-dotenv==1.0.0
groq==0.4.2
pydantic==2.5.2
requests==2.31.0
beautifulsoup4==4.12.2
pdf2image==1.16.3
